// src/app/shared/models/VeiculoSelecionado.model.ts
export class VeiculoSelecionado{
    responsavelNome: string;
    iotSituacao: string;
    situacao: string;
    responsavelTelefone: string;
    placa: string;
    latitude: number;
    responsavelId: any;
    longitude: number;
    nome: string;
    id: number;
    iotSituacaoId:number;
    veiculoId: number;
    violacaoCarenagem: boolean;
    violacaoBlindagem:boolean;
    locatarioId:number;
    veiculoSuporteId:number;
    apelido:string;
    veiculoSuportePlaca:string;
    suporteRoubo:boolean;
    operando: boolean;
    suporteManutencao: boolean;
    suporteRecolhimento: boolean;
    suporteReposicao: boolean;
    localizadorData:string;
    output1: boolean;
    output2: boolean;
    bateriaPrincipal: number;
}